refactor: extract ALE runner source CLI commands
