+++ examples/cli-command-module-size-ownership-command-modularization-smoke.py
@@ -10,7 +10,7 @@ CLI_COMMANDS = ROOT / "loopx" / "cli_commands"
 
 DEFAULT_MAX_LINES = 500
 LEGACY_MODULE_LIMITS = {
-    "agents_last_exam.py": 1850,
+    "agents_last_exam.py": 1560,
     "benchmark_review_lifecycle.py": 1300,
     "benchmark_run_ledger.py": 1360,
     "history.py": 720,
