+++ loopx/cli_commands/__init__.py
@@ -16,6 +16,10 @@ from .agents_last_exam_local_plan import (
     handle_agents_last_exam_local_plan_command,
     register_agents_last_exam_local_plan_commands,
 )
+from .agents_last_exam_runner_source import (
+    handle_agents_last_exam_runner_source_command,
+    register_agents_last_exam_runner_source_commands,
+)
 from .agentissue_runner_flow import (
     handle_agentissue_runner_flow_command,
     register_agentissue_runner_flow_commands,
@@ -120,6 +124,7 @@ from .worker_bridge import handle_worker_bridge_command, register_worker_bridge_
 __all__ = [
     "handle_agents_last_exam_command",
     "handle_agents_last_exam_local_plan_command",
+    "handle_agents_last_exam_runner_source_command",
     "handle_agentissue_runner_flow_command",
     "handle_benchmark_boundary_command",
     "handle_benchmark_command",
@@ -168,6 +173,7 @@ __all__ = [
     "handle_worker_bridge_command",
     "register_agents_last_exam_commands",
     "register_agents_last_exam_local_plan_commands",
+    "register_agents_last_exam_runner_source_commands",
     "register_agentissue_runner_flow_commands",
     "register_benchmark_boundary_commands",
     "register_benchmark_command_group",
